+++ CHANGELOG.md
@@ -27,11 +27,13 @@ Anticipated release: July 28, 2021
 - prevented some unexpected JSON parsing errors from being escalated to the error log. ([#3176])
 - Reduced the size of the JWT payload because sys admins had a 24KB payload and the max size is 4KB. ([#3246])
 - fixed error for users with view-document permission not being able to load an APD([#3264])
+- added audit log to track changes to user access ([#3262])
 - remove sys admins being returned in affiliations ([#3295])
 - added end-to-end test for APD overview section ([#3282])
 - added end-to-end tests for filling out key personnel page ([#3284])
 - added end-to-end tests for previous activities page ([#3312])
 - fixed export view of activity budgets ([#3388])
+- added end-to-end tests for assurances and compliance page ([#3390])
 - added end-to-end tests for activity schedule summary page ([#3391])
 
 # Previous releases
@@ -56,5 +58,7 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 [#3282]: https://github.com/CMSgov/eAPD/issues/3282
 [#3284]: https://github.com/CMSgov/eAPD/issues/3284
 [#3312]: https://github.com/CMSgov/eAPD/issues/3312
+[#3262]: https://github.com/CMSgov/eAPD/issues/3262
 [#3388]: https://github.com/CMSgov/eAPD/issues/3388
+[#3390]: https://github.com/CMSgov/eAPD/issues/3390
 [#3391]: https://github.com/CMSgov/eAPD/issues/3391